(original rev 948) pushing in new dataextents
git-svn-id: file:///tmp/current/omni/branches/staging@729 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/system/omCacheManager.cpp
@@ -126,7 +126,7 @@ void OmCacheManager::CleanCacheGroup(OmCacheGroup group)
 
 	//compute target size for group
 
-	int32_t target_size = (mTargetRatio) * mCacheMap[group].MaxSize;
+	unsigned long target_size = (mTargetRatio) * mCacheMap[group].MaxSize;
 
 	//get caches in group
 	set < OmCacheBase * >&cache_set = mCacheMap[group].CacheSet;
